made some adjustments to these classes

+++ ecommerce/src/main/java/com/example/ecommerce/dto/checkout/CheckoutItemDto.java
@@ -2,14 +2,14 @@ package com.example.ecommerce.dto.checkout;
 
 public class CheckoutItemDto {
     private String productName;
-    private int  quantity;
+    private Integer  quantity;
     private double price;
-    private long productId;
+    private Integer productId;
 
     public CheckoutItemDto() {
     }
 
-    public CheckoutItemDto(String productName, int quantity, double price, long productId) {
+    public CheckoutItemDto(String productName, int quantity, double price, Integer productId) {
         this.productName = productName;
         this.quantity = quantity;
         this.price = price;
@@ -28,7 +28,7 @@ public class CheckoutItemDto {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(Integer quantity) {
         this.quantity = quantity;
     }
 
@@ -44,7 +44,7 @@ public class CheckoutItemDto {
         return productId;
     }
 
-    public void setProductId(long productId) {
+    public void setProductId(Integer productId) {
         this.productId = productId;
     }
 }
